Task #85, change max_km param to float

+++ python-flask-reactjs-map/geo_clustering/c/distances_from_points_grouping.c
@@ -24,7 +24,7 @@ int make_distances_with_max(
     indent_t indent,
     const geo_clustered_point_t *const points,
     uint32_t num_points,
-    const int max_km,
+    const float max_km,
     scratch_buf_t *distances_scratch_buf
 ) {
     scratch_buf_t scratch_buf;
